Add intelligent tiering, only upload local files
Intelligent tiering support was added in ad979a6bead960c2058bacf818a3ee51a9153bcf, we need to support it in the upload script. Also we don't need to upload remote files, it's a waste of storage and money. We can always retrieve them from the servers of origin. I'm not sure if we even need to index them to the database. If there's a strong reason for that I might add an option to exclude/include the remote files from upload.

Signed-off-by: Sergey Shpikin <rkfg@rkfg.me>